requirements.txt
pydantic
pytest
pyyaml
ruff